Refactor + fixes
