+++ src/graphql/JobApplication/Types/GraphQLJobApplication.ts
@@ -1,6 +1,5 @@
 import { GraphQLObjectType } from "graphql";
 import { nonNull, String } from "../../fieldTypes";
-import { GraphQLOffer } from "../../Offer/Types/GraphQLOffer";
 
 const GraphQLJobApplication = new GraphQLObjectType({
   name: "JobApplication",
@@ -8,8 +7,8 @@ const GraphQLJobApplication = new GraphQLObjectType({
     applicantUuid: {
       type: nonNull(String)
     },
-    offer: {
-      type: nonNull(GraphQLOffer)
+    offerUuid: {
+      type: nonNull(String)
     }
   })
 });
