+++ src/graphql/User/queries.ts
@@ -1,11 +1,11 @@
 import { GraphQLUser } from "./Types/GraphQLUser";
-import { IApolloServerContext } from "../../server";
+import { IApolloServerContext } from "../../graphqlContext";
 import { UserRepository } from "../../models/User/Repository";
 
 export const userQueries = {
   me: {
     type: GraphQLUser,
     resolve: (_: undefined, __: object, { currentUser }: IApolloServerContext) =>
-      UserRepository.findByEmail(currentUser!.email)
+      UserRepository.findByEmail(currentUser.email)
   }
 };
